[tests][config] Move tests dedicated to the seiko plugin to a specific file
