+++ tests/test_config_parser.py
@@ -264,64 +264,3 @@ def test_specific_settings(sample_config, expected_settings):
             assert sample_config["parallel_printer"][k] == v, f"Fault key: {k}"
             continue
         assert sample_config["misc"][k] == v, f"Fault key: {k}"
-
-
-@pytest.mark.parametrize(
-    "sample_config,expected_settings",
-    [
-        # Config with user settings vs expected parsed settings
-        (
-            # default-settings
-            """
-            [misc]
-            emulation=seiko-qt2100
-            [parallel_printer]
-            [serial_printer]
-            """,
-            {
-                "enable-csv": "yes",
-                "enable-graph": "yes",
-                "vertical": "yes",
-                "cutoff": "true",
-            },
-        ),
-        (
-            # edited-settings
-            """
-            [misc]
-            emulation=seiko-qt2100
-            [parallel_printer]
-            [serial_printer]
-            [seiko-qt2100]
-            enable-graph=xxx
-            cutoff=false
-            vertical=no
-            """,
-            {
-                "enable-csv": "yes",
-                "enable-graph": "no",  # garbage fixed
-                "vertical": "no",
-                "cutoff": "false",
-            },
-        ),
-        (
-            # numeric-cutoff
-            """
-            [misc]
-            emulation=seiko-qt2100
-            [parallel_printer]
-            [serial_printer]
-            [seiko-qt2100]
-            cutoff=1.0
-            """,
-            {
-                "cutoff": "1.0",  # tristate variable parsed (bool, float, int)
-            },
-        ),
-    ],
-    ids=["default-settings", "edited-settings", "numeric-cutoff"],
-    indirect=["sample_config"],  # Send sample_config val to the fixture
-)
-def test_seiko_default_settings(sample_config, expected_settings):
-    for k, v in expected_settings.items():
-        assert sample_config["seiko-qt2100"][k] == v, f"Fault key: {k}"
